initial multi-gateway support
This commit allows for coordination of the target state within all
gateways in a gateway group. This functionality is only supported when
all requests are sent to a single gateway. Sending requests to mutliple
gateways within the group may result in error if a gateway receives a
new request while in an outdated state. With the state_update_notify
config option, gateways register a watch on the group's OMAP object and
send notify signals to the object after state changes. Gateways check
for OMAP updates upon receiving the watch callback as well as
periodically according to an state_update_interval_sec config parameter.

Signed-off-by: Sandy Kaur <sandy.kaur@ibm.com>